refactor(artifacts): extract download service
